// src/types.ts
export type ParsedArgsValueT = {
  outDir?: string | undefined;
  inDir?: string | undefined;
  funcName?: string | undefined;
  args?: string[] | undefined;
  port?: string | undefined;
  help?: boolean | undefined;
  list?: boolean | undefined;
  update?: boolean | undefined;
  version?: boolean | undefined;
  remove?: boolean | undefined;
}